+++ src/main.py
@@ -3,6 +3,8 @@ from PyQt5.QtCore import *
 from PyQt5.QtGui import QFont, QPixmap
 from PyQt5.QtWidgets import *
 from DirectoryManager import DirectoryManager
+
+from Controller.ConfigurationRecordController import ConfigurationRecordController
 from GUI.FirstTimeWindow import FirstTimeWindow
 
 
@@ -10,9 +12,17 @@ if __name__ == '__main__':
     # Create QApplication and QFrame instance
     app = QApplication([])
 
-    # Show the window and run the application
-    window = FirstTimeWindow()
-    window.show()
+    # Try open SQLite configuration
+    configuration_database = ConfigurationRecordController()
+    if not configuration_database.connect_to_database():
+        # Show the first time window
+        window = FirstTimeWindow()
+        window.show()
+    else:
+        # TODO: set program settings based on database contents
+        print("Opened configuration file.")
+        exit()
+
     app.exec()